% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition.R
\name{add_agent}
\alias{add_agent}
\title{Create an agent}
\usage{
add_agent(
  model,
  goal,
  group_number = 1,
  standing_start = 0.1,
  individual_differences = FALSE
)
}
\description{
Create an agent
}
