% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition.R
\name{add_group}
\alias{add_group}
\title{Add a group of agents to the simulation}
\usage{
add_group(
  model,
  goal,
  agent_number = 1,
  group_number = 1,
  standing_start = 0.1,
  individual_differences = FALSE
)
}
\arguments{
\item{agent_number}{Numeric denoting the number of agents to add. Defaults
to \code{1}.}

\item{standing_start}{Numeric denoting the factor of their preferred speed
that agents move when they just came from standing still. Defaults to
\code{0.1}.}

\item{individual_differences}{Logical denoting whether to use the standard
deviations in the parameter list to create some variation in the parameters.
Defaults to \code{FALSE}.}

\item{object}{Object of the \code{\link[predped]{predped-class}}.}
}
\value{
List of instances of the \code{\link[predped]{agent-class}}.
}
\description{
This function is the same as the one defined in \code{predped}, but with the 
exception that it uses the \code{add_agent} function defined in this same 
file instead.
}
