% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tablet.R
\name{assign_tablet}
\alias{assign_tablet}
\title{Assign a tablet to a participant}
\usage{
assign_tablet(state)
}
\arguments{
\item{state}{Object of the \code{\link[predped]{state-class}} containing the
current state.}
}
\description{
Draws a quasi-random tablet from the list of free tablets for each agent that
currently does not have a goal. When a "sit" tablet is free, the function
will check whether its the agents turn to sit down. If not, then we assign a
"walk" tablet. If it is the agents' turn, then we assign the agent a "sit"
tablet instead. If no "sit" tablets are free, then the agent is automatically
assigned a "walk" tablet.
}
\details{
Basically a function that does the bookkeeping of the experiment.
}
