% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{goal_stack}
\alias{goal_stack}
\title{Create the goal stack}
\usage{
goal_stack(room)
}
\arguments{
\item{room}{Object of the \code{\link[predped]{background-class}} containing 
the experimental room. This function assumes that all objects with "tablet" in
the name are to be walking goals, while those with "sofa" in the name are 
to be the sitting goals.}
}
\description{
Create the goal stack
}
