% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition.R
\name{initial_condition}
\alias{initial_condition}
\title{Create an initial condition}
\usage{
initial_condition(
  model,
  N,
  goal_stack,
  walk_duration,
  sit_duration,
  group_size = matrix(c(1, 1), nrow = 2),
  ...
)
}
\description{
Create an initial condition
}
