% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Create a function that will simulate a given experimental procedure with 
the following parameters.}
\usage{
simulate(
  model,
  iterations = 1800,
  N = c(6, 2),
  cycle_duration = 1800,
  walk_duration = function(x) abs(rnorm(1, 10, 2)),
  fx = function(x) x,
  time_step = 0.5,
  max_agents = NULL,
  group_size = matrix(c(1, 1), nrow = 1),
  individual_differences = FALSE,
  standing_start = 0.1,
  ...
)
}
\arguments{
\item{model}{Object of the \code{\link[predped]{predped-class}} specifying 
the details of the simulation.}

\item{iterations}{Integer denoting the number of iterations to run. Defaults 
to \code{1800}.}

\item{N}{Vector of integers denoting the number of participants that should 
walk around (first number) and the number of participants that should sit 
down (second number). The sample size is limited by the number of tablets in 
the environment (by default 8 "walk" and 2 "sit" tablets). Defaults to 6 
walking agents and 2 sitting ones.}

\item{cycle_duration}{Integer denoting how long a single cycle of the 
experiment takes. During each cycle, every participant should sit down a 
specified amount of time and walk for the rest of the cycle. Defaults to 
\code{1800}, or 900sec if \code{time_step = 0.5}.}

\item{walk_duration}{Integer or function that determines the duration 
of each goal assigned to a "walk" tablet in the square. Defaults to a normal
distribution with mean 10 and standard deviation 2.}

\item{fx}{Function that takes in and returns an object of the
\code{\link[predped]{state-class}}. This will be executed at the beginning of 
each iteration and allows users some flexibility in their simulations. For 
example useful when simulating evacuations (giving everyone "goal exit") or 
trying to guide behavior in any other way. Defaults to "\(x) x", meaning the 
state remains unaltered.}

\item{time_step}{Numeric denoting the number of seconds each discrete step in
time should mimic. Defaults to \code{0.5}, or half a second.}

\item{...}{Additional arguments that control the simulation (see 
\code{\link[predped]{simulate,predped-method}})}
}
\value{
Trace, or list of different states at the different iterations.
}
\description{
Create a function that will simulate a given experimental procedure with 
the following parameters.
}
